# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Feiticeiros e Maldições – RPG Character Sheet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bosses:
    get:
      operationId: listBosses
      tags: [bosses]
      summary: List bosses
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BossSummary"
    post:
      operationId: createBoss
      tags: [bosses]
      summary: Create boss
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BossInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boss"
  /bosses/{id}:
    get:
      operationId: getBoss
      tags: [bosses]
      summary: Get boss
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Boss
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boss"
    patch:
      operationId: updateBoss
      tags: [bosses]
      summary: Update boss
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BossUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boss"
    delete:
      operationId: deleteBoss
      tags: [bosses]
      summary: Delete boss
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignSummary"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    BossSummary:
      type: object
      required: [id, name, vd, category, size, grade, hp, maxHp, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        photoUrl: { type: ["string", "null"] }
        vd: { type: integer }
        category: { type: string }
        size: { type: string }
        grade: { type: string }
        hp: { type: integer }
        maxHp: { type: integer }
        innateTechnique: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    Boss:
      type: object
      required:
        - id
        - name
        - vd
        - category
        - size
        - grade
        - strength
        - dexterity
        - constitution
        - intelligence
        - wisdom
        - charisma
        - hp
        - maxHp
        - energy
        - maxEnergy
        - armorClass
        - attention
        - movement
        - hitDice
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        name: { type: string }
        photoUrl: { type: ["string", "null"] }
        vd: { type: integer }
        category: { type: string }
        size: { type: string }
        grade: { type: string }
        domain: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        appearance: { type: ["string", "null"] }
        strength: { type: integer }
        dexterity: { type: integer }
        constitution: { type: integer }
        intelligence: { type: integer }
        wisdom: { type: integer }
        charisma: { type: integer }
        hp: { type: integer }
        maxHp: { type: integer }
        energy: { type: integer }
        maxEnergy: { type: integer }
        armorClass: { type: integer }
        attention: { type: integer }
        movement: { type: string }
        hitDice: { type: string }
        abilities: { type: ["string", "null"] }
        innateTechnique: { type: ["string", "null"] }
        techniqueDescription: { type: ["string", "null"] }
        weaknesses: { type: ["string", "null"] }
        resistances: { type: ["string", "null"] }
        loot: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    BossInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        photoUrl: { type: string }
        vd: { type: integer, default: 10 }
        category: { type: string, default: "Maldição" }
        size: { type: string, default: "Médio" }
        grade: { type: string, default: "3° Grau" }
        domain: { type: string }
        description: { type: string }
        appearance: { type: string }
        strength: { type: integer }
        dexterity: { type: integer }
        constitution: { type: integer }
        intelligence: { type: integer }
        wisdom: { type: integer }
        charisma: { type: integer }
        hp: { type: integer }
        maxHp: { type: integer }
        energy: { type: integer }
        maxEnergy: { type: integer }
        armorClass: { type: integer }
        attention: { type: integer }
        movement: { type: string }
        hitDice: { type: string }
        abilities: { type: string }
        innateTechnique: { type: string }
        techniqueDescription: { type: string }
        weaknesses: { type: string }
        resistances: { type: string }
        loot: { type: string }
        notes: { type: string }
    BossUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        photoUrl: { type: string }
        vd: { type: integer }
        category: { type: string }
        size: { type: string }
        grade: { type: string }
        domain: { type: string }
        description: { type: string }
        appearance: { type: string }
        strength: { type: integer }
        dexterity: { type: integer }
        constitution: { type: integer }
        intelligence: { type: integer }
        wisdom: { type: integer }
        charisma: { type: integer }
        hp: { type: integer }
        maxHp: { type: integer }
        energy: { type: integer }
        maxEnergy: { type: integer }
        armorClass: { type: integer }
        attention: { type: integer }
        movement: { type: string }
        hitDice: { type: string }
        abilities: { type: string }
        innateTechnique: { type: string }
        techniqueDescription: { type: string }
        weaknesses: { type: string }
        resistances: { type: string }
        loot: { type: string }
        notes: { type: string }
    CampaignSummary:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        coverUrl: { type: ["string", "null"] }
        status: { type: string }
        currentArc: { type: ["string", "null"] }
        partyName: { type: ["string", "null"] }
        synopsis: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    Campaign:
      type: object
      required: [id, name, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        coverUrl: { type: ["string", "null"] }
        synopsis: { type: ["string", "null"] }
        setting: { type: ["string", "null"] }
        status: { type: string }
        currentArc: { type: ["string", "null"] }
        nextSession: { type: ["string", "null"] }
        partyName: { type: ["string", "null"] }
        playerCharacterIds: { type: ["string", "null"] }
        bossIds: { type: ["string", "null"] }
        sessionLog: { type: ["string", "null"] }
        npcs: { type: ["string", "null"] }
        locations: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CampaignInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        coverUrl: { type: string }
        synopsis: { type: string }
        setting: { type: string }
        status: { type: string, default: "Em andamento" }
        currentArc: { type: string }
        nextSession: { type: string }
        partyName: { type: string }
        playerCharacterIds: { type: string }
        bossIds: { type: string }
        sessionLog: { type: string }
        npcs: { type: string }
        locations: { type: string }
        notes: { type: string }
    CampaignUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        coverUrl: { type: string }
        synopsis: { type: string }
        setting: { type: string }
        status: { type: string }
        currentArc: { type: string }
        nextSession: { type: string }
        partyName: { type: string }
        playerCharacterIds: { type: string }
        bossIds: { type: string }
        sessionLog: { type: string }
        npcs: { type: string }
        locations: { type: string }
        notes: { type: string }
...