# lib/api-zod/src/generated/types/bossInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Feiticeiros e Maldições – RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface BossInput {
  /** @minLength 1 */
  name: string;
  photoUrl?: string;
  vd?: number;
  category?: string;
  size?: string;
  grade?: string;
  domain?: string;
  description?: string;
  appearance?: string;
  strength?: number;
  dexterity?: number;
  constitution?: number;
  intelligence?: number;
  wisdom?: number;
  charisma?: number;
  hp?: number;
  maxHp?: number;
  energy?: number;
  maxEnergy?: number;
  armorClass?: number;
  attention?: number;
  movement?: string;
  hitDice?: string;
  abilities?: string;
  innateTechnique?: string;
  techniqueDescription?: string;
  weaknesses?: string;
  resistances?: string;
  loot?: string;
  notes?: string;
}