# lib/api-zod/src/generated/types/bossSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Feiticeiros e Maldições – RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface BossSummary {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  vd: number;
  category: string;
  size: string;
  grade: string;
  hp: number;
  maxHp: number;
  /** @nullable */
  innateTechnique?: string | null;
  createdAt: Date;
}