# lib/api-zod/src/generated/types/campaignSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Feiticeiros e Maldições – RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignSummary {
  id: number;
  name: string;
  /** @nullable */
  coverUrl?: string | null;
  status: string;
  /** @nullable */
  currentArc?: string | null;
  /** @nullable */
  partyName?: string | null;
  /** @nullable */
  synopsis?: string | null;
  createdAt: Date;
}